--- packages/bundle/src/boot/actual/full-es5.ts
@@ -1,36 +1,0 @@
-import defaultCreateDirectLine from '../../createDirectLine';
-import defaultCreateDirectLineAppServiceExtension from '../../createDirectLineAppServiceExtension';
-import buildInfo from '../../buildInfo';
-import ReactWebChat from './full';
-
-buildInfo.set('variant', 'full-es5');
-
-const createDirectLine = (options: Omit<Parameters<typeof defaultCreateDirectLine>[0], 'botAgent'>) => {
-  (options as any).botAgent &&
-    console.warn(
-      'Web Chat: Developers are not currently allowed to set botAgent. See https://github.com/microsoft/BotFramework-WebChat/issues/2119 for more details.'
-    );
-
-  return defaultCreateDirectLine({ ...options, botAgent: `WebChat/${buildInfo.version} (ES5)` });
-};
-
-const createDirectLineAppServiceExtension = (
-  options: Omit<Parameters<typeof defaultCreateDirectLineAppServiceExtension>[0], 'botAgent'>
-) => {
-  (options as any).botAgent &&
-    console.warn(
-      'Web Chat: Developers are not currently allowed to set botAgent. See https://github.com/microsoft/BotFramework-WebChat/issues/2119 for more details.'
-    );
-
-  return defaultCreateDirectLineAppServiceExtension({ ...options, botAgent: `WebChat/${buildInfo.version} (ES5)` });
-};
-
-// #region Re-exports
-export * from './full';
-// #endregion
-
-// #region Local exports
-export default ReactWebChat;
-
-export { createDirectLine, createDirectLineAppServiceExtension };
-// #endregion
