+++ packages/bundle/src/boot/actual/minimal.ts
@@ -4,35 +4,15 @@ import * as componentDecorator from 'botframework-webchat-component/decorator';
 import * as internal from 'botframework-webchat-component/internal';
 
 import buildInfo from '../../buildInfo';
-import defaultCreateDirectLine from '../../createDirectLine';
-import defaultCreateDirectLineAppServiceExtension from '../../createDirectLineAppServiceExtension';
 import coreRenderWebChat from '../../renderWebChat';
+import createDirectLine from './private/createDirectLine';
+import createDirectLineAppServiceExtension from './private/createDirectLineAppServiceExtension';
 
 buildInfo.set('variant', 'minimal');
 
 const { object: buildInfoObject, version } = buildInfo;
 
 const renderWebChat = coreRenderWebChat.bind(null, ReactWebChat);
-
-const createDirectLine = (options: Omit<Parameters<typeof defaultCreateDirectLine>[0], 'botAgent'>) => {
-  (options as any).botAgent &&
-    console.warn(
-      'Web Chat: Developers are not currently allowed to set botAgent. See https://github.com/microsoft/BotFramework-WebChat/issues/2119 for more details.'
-    );
-
-  return defaultCreateDirectLine({ ...options, botAgent: `WebChat/${version} (Minimal)` });
-};
-
-const createDirectLineAppServiceExtension = (
-  options: Omit<Parameters<typeof defaultCreateDirectLineAppServiceExtension>[0], 'botAgent'>
-) => {
-  (options as any).botAgent &&
-    console.warn(
-      'Web Chat: Developers are not currently allowed to set botAgent. See https://github.com/microsoft/BotFramework-WebChat/issues/2119 for more details.'
-    );
-
-  return defaultCreateDirectLineAppServiceExtension({ ...options, botAgent: `WebChat/${version} (Minimal)` });
-};
 
 const decorator = Object.freeze({
   ...apiDecorator,
