+++ packages/bundle/src/boot/bundle/full-es5.ts
@@ -2,15 +2,8 @@
 import './polyfill/es5';
 
 import buildInfo from '../../buildInfo';
-import * as actual from '../actual/full-es5';
-import * as middleware from '../actual/middleware';
 
-// Until we have a development-specific bundle, we are not shipping createStoreWithDevTools in bundle.
-const { createStoreWithDevTools: _createStoreWithDevTools, ...exports } = actual;
+buildInfo.set('variant', 'full-es5');
 
-window['WebChat'] = Object.freeze({
-  ...window['WebChat'],
-  ...exports,
-  buildInfo: buildInfo.object,
-  middleware
-});
+// "./full" is IIFE, just importing it will inject to globalThis.
+import './full';
