--- packages/bundle/src/boot/exports/full-es5.ts
@@ -1,4 +1,0 @@
-import ReactWebChat from '../actual/full-es5';
-
-export * from '../actual/full-es5';
-export default ReactWebChat;
