+++ packages/bundle/src/boot/exports/index.ts
@@ -1,3 +1,4 @@
+// Default is not re-exported when using `export *`.
 import ReactWebChat from '../actual/full';
 
 export * from '../actual/full';
