--- packages/bundle/src/boot/exports/minimal.ts
@@ -1,4 +1,0 @@
-import ReactWebChat from '../actual/minimal';
-
-export * from '../actual/minimal';
-export default ReactWebChat;
